Changed from Interface to Abstract Class
GlobalLootModifierSerializer is now an abstract class and the abstract implementation in LootModifier was no longer needed.
